Use MapStruct to convert between spec-grpc and spec classes.
The spec-grpc classes are ugly, and we can't let users use them.
When using MapStruct an effort has been made to not use 'manual'
conversion (i.e. a 'default' implementation of the Mapper interface methods).

When automatic conversion happens, when the spec-grpc and spec classes differ
a compile error will be raised. Which in turn helps us keep our manually
generated spec classes in sync with the spec-grpc ones.